requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0